More changes to the build system
