+++ test/package.json
@@ -1,17 +1,13 @@
 {
     "name": "test",
-    "tests": [
-        "stateMgr.js",
-        "msgChannel.js",
-        "lex.js",
-        "parse/lex.js",
-        "parse/basic.js",
-        "parse/lang.ts"
-    ],
+    "type": "module",
     "scripts": {
-        "build:all": "find $(cd .. ; pwd) -iname \\*.test.ts | xargs -d \\\\n -I {} pnpm --prefix ../ exec esbuild {} --outdir=build/test --bundle --sourcemap --platform=node --format=esm"
+        "postinstall": "$npm_execpath exec mkjson all"
     },
     "dependencies": {
         "@j-cake/jcake-utils": "link:../build"
+    },
+    "devDependencies": {
+        "@j-cake/mkjson": "latest"
     }
 }
