+++ tsconfig.json
@@ -15,16 +15,16 @@
         "forceConsistentCasingInFileNames": true,
         "importHelpers": true,
         "strict": true,
-        "baseUrl": "src",
+        "baseUrl": "./src",
         "paths": {
-            "#args": ["src/args/index.ts"],
-            "#buffer": ["src/buffer/index.ts"],
-            "#db": ["src/db/index.ts"],
-            "#io": ["src/io/index.ts"],
-            "#iter": ["src/iter/index.ts"],
-            "#msg": ["src/msg/index.ts"],
-            "#parse": ["src/parse/index.ts"],
-            "#state": ["src/state/index.ts"],
+            "#args": ["args/index.ts"],
+            "#buffer": ["buffer/index.ts"],
+            "#db": ["db/index.ts"],
+            "#io": ["io/index.ts"],
+            "#iter": ["iter/index.ts"],
+            "#msg": ["msg/index.ts"],
+            "#parse": ["parse/index.ts"],
+            "#state": ["state/index.ts"],
 
         }
     },
